backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
boto3==1.40.45
botocore==1.40.45
celery==5.3.4
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==4.4.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
kombu==5.5.4
moviepy==1.0.3
numpy==2.3.3
packaging==25.0
pillow==10.4.0
proglog==0.1.12
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
